requirements.txt
python==3.6.7
numpy==1.15.4
scipy==1.1.0
matplotlib==3.0.2
seaborn==0.9.0
pandas==0.24.2
networkx==2.2
pytorch==0.4.1
scikit-learn==0.20.2
tensorflow==1.12.0
tensorboard==1.12.0
tensorboardX==1.5
fastprogress==0.1.18